// src/1_pages/feedback/api/types/response.ts
export type FeedbackMyResponse = {
  id: number;
  type: string;
  page: string;
  rank: string;
  text_message: string;
  status: string;
  create_add: string;
};

export type FeedbackAllResponse = {
  id: number;
  type: string;
  page: string;
  rank: string;
  text_message: string;
  name_user: string;
  status: string;
  create_add: string;
};